portfolio/requirements.txt
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.9.0
asgiref==3.8.1
Authlib==1.6.0
beautifulsoup4==4.12.3
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
confection==0.1.5
cryptography==45.0.4
cymem==2.0.11
Deprecated==1.2.18
distro==1.9.0
Django==5.0.7
django-bootstrap4==24.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring_parser==0.16
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
et_xmlfile==2.0.0
factur-x==3.6
Faker==37.4.2
fastapi==0.115.13
filelock==3.18.0
fpdf==1.7.2
google-adk==1.4.2
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.97.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-dlp==3.30.0
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
livereload==2.7.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.3.0
openai==1.97.1
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.3.1
phonenumbers==8.13.55
pikepdf==9.10.2
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
presidio_analyzer==2.2.358
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.3
pypdf==5.9.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.1
srsly==2.5.1
sse-starlette==2.3.6
starlette==0.46.2
tenacity==8.5.0
thinc==8.3.6
tldextract==5.3.0
tornado==6.4.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.34.3
uvloop==0.21.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
whitenoise==6.9.0
wrapt==1.17.2
zipp==3.23.0
